refactor: move App.tsx to app/src/App.tsx - Moved App.tsx from components/ to src/ root - Updated imports in App.tsx and App.test.tsx
