+++ app/src/App.tsx
@@ -1,8 +1,8 @@
 // client/src/components/App.tsx
 import React, { useState, useEffect } from 'react';
-import Signup from './auth/Signup';
-import ProfileSetupScreen from '../screens/ProfileSetupScreen';
-import ProfileSelectionScreen from '../screens/ProfileSelectionScreen';
+import Signup from './components/auth/Signup';
+import ProfileSetupScreen from './screens/ProfileSetupScreen';
+import ProfileSelectionScreen from './screens/ProfileSelectionScreen';
 
 // Define interfaces for User and Child (adjust fields based on your API response)
 interface User {
