+++ app/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client'; 
-import App from './components/App';
+import App from './App';
 
 // Create a root and render the App component
 const root = ReactDOM.createRoot(document.getElementById('root'));
